fix bug

+++ model/dataset.py
@@ -84,7 +84,7 @@ class TweetDataset(Dataset):
         if self.transform:
             image = self.transform(image)
         index_of_one = torch.tensor([self.tag2label[x.strip()] for x in tags])
-        label = torch.zeros(len(self.tag2label)).long()
+        label = torch.zeros(len(self.tag2label))
         label[index_of_one] = 1
         text_preprocessed = self.text_vocab.preprocess(text)
         return self.text_vocab.tensorFromSentence(text_preprocessed), image, label
